Fix `start_offset` when playing from start

+++ src/engine/components/timeline/timestamp.rs
@@ -42,8 +42,8 @@ impl Timestamp {
         }
     }
     /// Like [`Timestamp::from_samples`], but rounds up instead of down.
-    pub const fn from_samples_ceil(samples: u64, sample_rate: u32, bpm_cents: u16) -> Self {
-        const fn div_ceil(a: u64, b: u64) -> u64 {
+    pub const fn from_samples_ceil(samples: usize, sample_rate: u32, bpm_cents: u16) -> Self {
+        const fn div_ceil(a: usize, b: usize) -> usize {
             let d = a / b;
             let r = a % b;
             if r != 0 {
@@ -54,8 +54,8 @@ impl Timestamp {
         }
 
         let beat_units = div_ceil(
-            samples * bpm_cents as u64 * UNITS_PER_BEAT as u64,
-            sample_rate as u64 * 60 * 100,
+            samples * bpm_cents as usize * UNITS_PER_BEAT as usize,
+            sample_rate as usize * 60 * 100,
         );
         Self {
             beat_units: beat_units as u32,
